refactor(wallet-drawer-content): rename component
