+++ components/index.js
@@ -73,5 +73,5 @@ export { default as ValidatedPaymentMethod } from './validated-payment-method';
 export { default as ValidatedSelect } from './validated-select';
 export { default as ValidatedStarRating } from './validated-star-rating.vue';
 export { default as ValidatedTextarea } from './validated-textarea.vue';
-export { default as WalletContent } from './wallet-content';
 export { default as WalletDrawer } from './wallet-drawer';
+export { default as WalletDrawerContent } from './wallet-drawer-content';
