+++ components/wallet-drawer-content.vue
